Big refactor: move all emails into seperate files, add snapshots
