+++ services/app-api/src/emailer/etaTemplates/newPackageCMSEmail.eta
